requirements.txt
crewai==0.80.0
GitPython==3.1.43
GitPython==3.1.43
pydantic==2.9.2
python_docx==1.1.2
PyYAML==6.0.1
PyYAML==6.0.2